This commit contains the follow changes: 1. Cancel job support in KafkaPushJob; 2. Topic cleanup in Parent Controller when kill job or job finishes; 3. Version creation in TopicMonitor only if it is master controller; 4. Improve the waiting logic in TestAdminConsumptionTask; 5. Not throw exception if there is no storage node for current killing job;
RB=842581
G=voldemort-reviewers
A=sdwu

+++ venice-common/src/main/java/com/linkedin/venice/kafka/TopicManager.java
@@ -74,9 +74,13 @@ public class TopicManager implements Closeable {
   }
 
   public void deleteTopic(String topicName) {
-    logger.info("Deleting topic: " + topicName);
-    // TODO: Stop using Kafka APIs which depend on ZK.
-    AdminUtils.deleteTopic(getZkUtils(), topicName);
+    if (listTopics().contains(topicName)) {
+      // TODO: Stop using Kafka APIs which depend on ZK.
+      logger.info("Deleting topic: " + topicName);
+      AdminUtils.deleteTopic(getZkUtils(), topicName);
+    } else {
+      logger.info("Topic: " +  topicName + " to be deleted doesn't exist");
+    }
   }
 
   public synchronized Set<String> listTopics() {
